requirement.txt
matplotlib==3.9.2
numpy==1.26.4
pillow==10.4.0
tifffile==2024.9.20
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchinfo==1.8.0
torchvision==0.18.1+cu118